requirements.txt
django==4.2.16
django-debug-toolbar==4.4.6
psycopg2-binary==2.9.9
requests==2.32.3
plotly==5.24.1
pandas==2.2.2
python-dotenv==1.0.1